// back/tests/rest/api/api.ts
export interface AddTask {
	/**
	 *
	 * @type {ScheduleLight}
	 * @memberof AddTask
	 */
	schedule: ScheduleLight;
	/**
	 *
	 * @type {TaskWorkList}
	 * @memberof AddTask
	 */
	work: TaskWorkList;
}
// ...
export interface Save {
	/**
	 *
	 * @type {string}
	 * @memberof Save
	 */
	type: SaveTypeEnum;
	/**
	 *
	 * @type {ConnectOptions}
	 * @memberof Save
	 */
	connectionInfo?: ConnectOptions;
	/**
	 *
	 * @type {string}
	 * @memberof Save
	 */
	path: string;
}

/**
 * @export
 * @enum {string}
 */
export enum SaveTypeEnum {
	Ssh = 'ssh',
	Local = 'local'
}
// ...
export interface ScheduleLight {
	/**
	 *
	 * @type {number}
	 * @memberof ScheduleLight
	 */
	interval: number;
}
// ...
export interface TaskOn {
	/**
	 *
	 * @type {ConnectOptions}
	 * @memberof TaskOn
	 */
	connectionInfo?: ConnectOptions;
	/**
	 *
	 * @type {string}
	 * @memberof TaskOn
	 */
	folder: string;
	/**
	 *
	 * @type {string}
	 * @memberof TaskOn
	 */
	type: TaskOnTypeEnum;
}

/**
 * @export
 * @enum {string}
 */
export enum TaskOnTypeEnum {
	Local = 'local',
	Ssh = 'ssh'
}

/**
 *
 * @export
 * @interface TaskWorkList
 */
export interface TaskWorkList {
	/**
	 *
	 * @type {string}
	 * @memberof TaskWorkList
	 */
	type: TaskWorkListTypeEnum;
	/**
	 *
	 * @type {TaskOn}
	 * @memberof TaskWorkList
	 */
	on: TaskOn;
	/**
	 *
	 * @type {Save}
	 * @memberof TaskWorkList
	 */
	save: Save;
}

/**
 * @export
 * @enum {string}
 */
export enum TaskWorkListTypeEnum {
	List = 'list'
}
// ...
export class TaskApi extends BaseAPI {
	/**
	 * Add a local config
	 * @param {AddTask} addTask
	 * @param {*} [options] Override http request option.
	 * @throws {RequiredError}
	 * @memberof TaskApi
	 */
	public taskCreateTask(addTask: AddTask, options?: any) {
		return TaskApiFp(this.configuration).taskCreateTask(addTask, options).then((request) => request(this.axios, this.basePath));
	}

	/**
	 * Get all configs
	 * @param {*} [options] Override http request option.
	 * @throws {RequiredError}
	 * @memberof TaskApi
	 */
	public taskGetConfig(options?: any) {
		return TaskApiFp(this.configuration).taskGetConfig(options).then((request) => request(this.axios, this.basePath));
	}

	/**
	 *
	 * @param {*} [options] Override http request option.
	 * @throws {RequiredError}
	 * @memberof TaskApi
	 */
	public taskRunAllTask(options?: any) {
		return TaskApiFp(this.configuration).taskRunAllTask(options).then((request) => request(this.axios, this.basePath));
	}

	/**
	 *
	 * @param {number} id
	 * @param {*} [options] Override http request option.
	 * @throws {RequiredError}
	 * @memberof TaskApi
	 */
	public taskRunTask(id: number, options?: any) {
		return TaskApiFp(this.configuration).taskRunTask(id, options).then((request) => request(this.axios, this.basePath));
	}

	/**
	 *
	 * @param {number} id
	 * @param {*} [options] Override http request option.
	 * @throws {RequiredError}
	 * @memberof TaskApi
	 */
	public taskStopTask(id: number, options?: any) {
		return TaskApiFp(this.configuration).taskStopTask(id, options).then((request) => request(this.axios, this.basePath));
	}
}